third_party/blink/web_tests/platform/generic/fast/forms/textarea/textarea-placeholder-dom-property-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=21248

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textarea.placeholder is ""
PASS textarea.placeholder is "p1"
PASS textarea.value is ""
PASS textarea.getAttribute("placeholder") is "p2"
PASS textarea.placeholder is "p2"
PASS successfullyParsed is true

TEST COMPLETE
